the second part of my mini Project

+++ lib/screens/home_screen_stream.dart
@@ -13,14 +13,14 @@ class HomeScreenStream extends StatefulWidget {
 
 class _HomeScreenStreamState extends State<HomeScreenStream> {
   bool isEditing = false;
+  bool isLoading = false;
   final noteController = TextEditingController();
-  void submit([Note? nota]) {
+  Future<void> submit([Note? nota]) async {
     final currentUserId = context.read<AuthProvider>().session?.user.id;
     if (currentUserId == null) {
       return;
     }
 
-    // update editing flag so the dialog updates correctly
     setState(() {
       isEditing = nota != null;
       if (isEditing) noteController.text = nota!.content;
@@ -40,22 +40,32 @@ class _HomeScreenStreamState extends State<HomeScreenStream> {
             child: Text("Cancel"),
           ),
           TextButton(
-            onPressed: () {
+            onPressed: () async {
+              //abre el dialog de carga cuando se presione la accion
+              isLoadingShow(isEditing ? "Actualizando..." : "Creando...");
+
               if (isEditing) {
                 final note = Note(
                   content: noteController.text.trim(),
                   id: nota!.id,
                   userId: nota.userId,
                 );
-                context.read<NoteProviderStream>().updateNote(note);
+
+                await context.read<NoteProviderStream>().updateNote(note);
               } else {
                 final note = Note(
                   content: noteController.text.trim(),
                   userId: currentUserId,
                 );
-                context.read<NoteProviderStream>().insertNote(note);
+
+                await context.read<NoteProviderStream>().insertNote(note);
               }
-              Navigator.pop(context);
+              //se cierran los dialogos cuando se terminan las acciones
+              if (mounted) {
+                Navigator.pop(context); // cierra loading
+                Navigator.pop(context); // cierra dialog
+              }
+
               noteController.clear();
             },
             child: Text("Aceptar"),
@@ -66,7 +76,51 @@ class _HomeScreenStreamState extends State<HomeScreenStream> {
   }
 
   Future<void> delete(int id) async {
-    await context.read<NoteProviderStream>().deleteNote(id);
+    showDialog(
+      context: context,
+      builder: (builder) => AlertDialog(
+        content: Text('Seguro que desea borrar?'),
+        actions: [
+          TextButton(
+            onPressed: () {
+              Navigator.pop(context);
+            },
+            child: Text('Cancelar'),
+          ),
+          TextButton(
+            onPressed: () async {
+              isLoadingShow('borrando....');
+              await context.read<NoteProviderStream>().deleteNote(id);
+
+              if (mounted) {
+                // ← Verifica que el widget siga vivo
+                Navigator.pop(context);
+                Navigator.pop(context);
+              }
+            },
+            child: Text('Aceptar'),
+          ),
+        ],
+      ),
+    );
+  }
+
+  void isLoadingShow(String accion) {
+    showDialog(
+      context: context,
+      barrierDismissible: false,
+      builder: (context) {
+        return AlertDialog(
+          content: Row(
+            children: [
+              const CircularProgressIndicator(),
+              const SizedBox(width: 20),
+              Expanded(child: Text(accion)),
+            ],
+          ),
+        );
+      },
+    );
   }
 
   /* Es para liberar memoria cuando se  elimina la pantalla */
@@ -79,16 +133,16 @@ class _HomeScreenStreamState extends State<HomeScreenStream> {
   @override
   Widget build(BuildContext context) {
     final size = MediaQuery.of(context).size;
-    final user_id = context.watch<AuthProvider>().session?.user.id;
+    final userId = context.watch<AuthProvider>().session?.user.id;
 
-    if (user_id == null) {
+    if (userId == null) {
       return const Scaffold(
         body: Center(child: Text('Necesitas iniciar sesión')),
       );
     }
 
     final noteProvider = context.watch<NoteProviderStream>();
-    final stream = noteProvider.list(user_id);
+    final stream = noteProvider.list(userId);
 
     return SafeArea(
       child: Scaffold(
@@ -116,49 +170,51 @@ class _HomeScreenStreamState extends State<HomeScreenStream> {
                   return CircularProgressIndicator();
                 } else {
                   var notas = snapshop.data!;
-                  
-                  return notas.isEmpty?Center(child: Text('No tienes notas'),):
-                  ListView.builder(
-                    itemCount: notas.length,
-                    itemBuilder: (BuildContext context, int index) {
-                      return Container(
-                        padding: EdgeInsets.symmetric(
-                          horizontal: 20,
-                          vertical: 12,
-                        ),
-                        margin: EdgeInsets.symmetric(
-                          horizontal: size.width * 0.1,
-                          vertical: size.height * 0.01,
-                        ),
-                        decoration: BoxDecoration(
-                          color: const Color.fromARGB(255, 97, 156, 99),
-                          borderRadius: BorderRadius.circular(10),
-                        ),
-                        child: Row(
-                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                          children: [
-                            Text(notas[index].content),
-                            Row(
-                              children: [
-                                IconButton(
-                                  onPressed: () async {
-                                    await delete(notas[index].id!);
-                                  },
-                                  icon: Icon(Icons.delete),
-                                ),
-                                IconButton(
-                                  onPressed: () {
-                                    submit(notas[index]);
-                                  },
-                                  icon: Icon(Icons.edit),
-                                ),
-                              ],
-                            ),
-                          ],
-                        ),
-                      );
-                    },
-                  );
+
+                  return notas.isEmpty
+                      ? Center(child: Text('No tienes notas'))
+                      : ListView.builder(
+                          itemCount: notas.length,
+                          itemBuilder: (BuildContext context, int index) {
+                            return Container(
+                              padding: EdgeInsets.symmetric(
+                                horizontal: 20,
+                                vertical: 12,
+                              ),
+                              margin: EdgeInsets.symmetric(
+                                horizontal: size.width * 0.1,
+                                vertical: size.height * 0.01,
+                              ),
+                              decoration: BoxDecoration(
+                                color: const Color.fromARGB(255, 97, 156, 99),
+                                borderRadius: BorderRadius.circular(10),
+                              ),
+                              child: Row(
+                                mainAxisAlignment:
+                                    MainAxisAlignment.spaceBetween,
+                                children: [
+                                  Text(notas[index].content),
+                                  Row(
+                                    children: [
+                                      IconButton(
+                                        onPressed: () {
+                                          delete(notas[index].id!);
+                                        },
+                                        icon: Icon(Icons.delete),
+                                      ),
+                                      IconButton(
+                                        onPressed: () {
+                                          submit(notas[index]);
+                                        },
+                                        icon: Icon(Icons.edit),
+                                      ),
+                                    ],
+                                  ),
+                                ],
+                              ),
+                            );
+                          },
+                        );
                 }
               },
             ),
